% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input.R
\name{check_input}
\alias{check_input}
\title{Check input data}
\usage{
check_input(x, depth = NULL)
}
\arguments{
\item{x}{vector of the variable of interest.}

\item{depth}{vector of depths at which \code{x} is measured.}
}
\value{
TRUE when the data is ok, FALSE when it is fully missing, FALSE and throws an error when \code{depth} is not the same length as \code{x}.
}
\description{
Check that the input data is usable as a CTD cast (not fully missing and of compatible length). Used by other functions.
}
\examples{
check_input(c(22.1, 23.4, 23.2), c(1, 2, 3))
check_input(c(NA, NA, NA), c(1, 2, 3))
\dontrun{
check_input(c(22.1, 23.4, 23.2), c(1, 2))}
}
