% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clined.R
\name{clined}
\alias{clined}
\title{Detect the position of a cline}
\usage{
clined(x, depth = NULL, k = 2, n.smooth = 0, ...)
}
\arguments{
\item{x}{vector of data (assumed to be approximately evenly spaced for moving window statistics to make sense).}

\item{k}{order of the window; the window size is 2k+1.}

\item{n.smooth}{integer, number of times to smooth the data before computing the moving standard deviation.}

\item{...}{not used.}
}
\value{
When \code{depth} is \code{NULL}, return the index of the cline, \code{i}; when \code{depth} is provided, return \code{depth[i]}, the value of the depth of the cline.
}
\description{
The cline of a variable is the depth at which it varies the most along the cast; here it is computed as the maximum of the moving standard deviation over the (possibly smoothed) cast.
}
\examples{
plot(-depth ~ temp, data=d, type="l")
thermocline <- clined(d$temp, d$depth)
abline(h=-thermocline, col="red")

plot(-depth ~ sal, data=d, type="l")
halocline <- clined(d$sal, d$depth)
abline(h=-halocline, col="red")
}
