% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest.R
\name{closest}
\alias{closest}
\alias{which.closest}
\title{Find closest matches}
\usage{
closest(from, within, outside = TRUE)

which.closest(from, within, outside = TRUE)
}
\arguments{
\item{from}{a numeric vector (or coercible as such); the values to be matched.}

\item{within}{a vector of the same nature as \code{from}; the values to be matched against.}

\item{outside}{logicial. When \code{TRUE} (the default) return the extremities for elements outside of the range \code{[min(within),max(within)]}. When \code{FALSE} return \code{NA} for such elements}
}
\value{
A vector of integers of the same length as \code{from}: for \code{closest()}, the value of the closest elements of \code{within}; for \code{which.closest()} the indexes of those values.
}
\description{
For each element of \code{from}, find the the closest element of \code{within}.
}
\examples{
set.seed(1)
to_match  <- round(runif(5, min=0, max=100), 1)
reference <- round(runif(10, min=20, max=80), 1)

# compare the vectors
to_match
reference

# get closest mactches, or their index
closest(to_match, within=reference)
which.closest(to_match, within=reference)

# verify graphically
plot(
  x=c(reference, to_match),
  y=c(rep(2, 10), rep(1, 5)), ylim=c(0, 3), yaxt="n",
  xlab="", ylab=""
)
axis(2, at=1:2, labels=c("to match", "reference"))
# join closest points
segments(
  x0=to_match, y0=1,
  x1=closest(from=to_match, within=reference), y1=2
)
# join closest points inside "within"
abline(v=range(reference), lty="22")
segments(
  x0=to_match, y0=1,
  x1=closest(from=to_match, within=reference, outside=FALSE), y1=2,
col="red")
}
