% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castr.R
\docType{data}
\name{ctd}
\alias{ctd}
\title{Several CTD casts}
\format{A data frame with 3134 rows and 6 variables:
\describe{
\item{date}{date of sampling}
\item{depth}{depth, in m, positive towards the bottom}
\item{temp}{temperature, in degrees Celsius}
\item{sal}{salinity, in PSU}
\item{fluo}{fluorescence of Chlorophyll a, in mg/m^3}
\item{sigma}{potential density, in kg/m^3 - 1000}
}}
\usage{
ctd
}
\description{
A time series of 41 CTD casts collected at the same station over two years, identified by the date of sampling.
}
\examples{
head(ctd)
plot_casts <- function(formula, x) {
  plot(formula, data=x, type="n")
  for (day in unique(x$date)) {
    lines(formula, data=subset(x, date==day))
  }
}
plot_casts(-depth ~ temp, ctd)
plot_casts(-depth ~ sal, ctd)
plot_casts(-depth ~ fluo, ctd)
plot_casts(-depth ~ sigma, ctd)
}
\keyword{datasets}
