% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castr.R
\docType{data}
\name{d}
\alias{d}
\title{A CTD cast}
\format{A data frame with 77 rows and 5 variables:
\describe{
\item{depth}{depth, in m, positive towards the bottom}
\item{temp}{temperature, in degrees Celsius}
\item{sal}{salinity, in PSU}
\item{chla}{chlorophyll a concentration estimated through fluorescence, in mg/m^3}
\item{sigma}{potential density, in kg/m^3 - 1000}
}}
\usage{
d
}
\description{
A dataset representing a single CTD cast.
}
\examples{
head(d)
plot(-depth ~ temp,  data=d, type="b")
plot(-depth ~ sal,   data=d, type="b")
plot(-depth ~ chla,  data=d, type="b")
plot(-depth ~ sigma, data=d, type="b")
}
\keyword{datasets}
