% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euphoticd.R
\name{euphoticd}
\alias{euphoticd}
\title{Detect the lower limit of the euphotic zone}
\usage{
euphoticd(chla, depth, fit = c("piecewise-linear", "polynomial"))
}
\arguments{
\item{chla}{vector of chlorophyll concentration, in mg m-3.}

\item{depth}{vector of depths at which \code{chl} is measured.}

\item{fit}{method use to represent the relationship between Ze and integrated chlorophyll.}
}
\value{
The depth of the lower limit of the euphotic zone.
}
\description{
Detect the lower limit of the euphotic zone
}
\details{
The method is that of Morel and Berthon (1989) with the new parameter values from Morel and Maritorena (2001)
}
\examples{
plot(-depth ~ chla, data=d, type="l")
Ze <- euphoticd(chla, depth)
abline(h=-Ze, col="chartreuse4")
Ze <- euphoticd(chla, depth, fit="poly")
# -> same result
}
\references{
Morel and Berthon (1989). Surface pigments, algal biomass profiles, and potential production of the euphotic layer: Relationships reinvestigated in view of remote‐sensing applications. Limnology and Oceanography, 34(8), 1545-1562.

Morel and Maritorena (2001). Bio-optical properties of oceanic waters: A reappraisal. Journal of Geophysical Research: Oceans, 106(C4). \url{https://doi.org/10.1029/2000JC000319}
}
\seealso{
\code{\link[=smooth]{smooth()}} for smoothing.

Other functions computing remarkable depths: \code{\link{clined}},
  \code{\link{maxd}}, \code{\link{mld}}
}
\concept{functions computing remarkable depths}
