% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxd.R
\name{maxd}
\alias{maxd}
\title{Detect the position of the maximum}
\usage{
maxd(x, depth = NULL, k = 2, n.smooth = 1, ...)
}
\arguments{
\item{x}{vector of data (assumed to be approximately evenly spaced for moving window statistics to make sense).}

\item{k}{order of the window; the window size is 2k+1.}

\item{n.smooth}{integer, number of times to smooth the data before detecting the maximum (1 by default, because we are rarely interested in the absolute maximum, which often corresponds to a spike, but rather in the region of the maximum values).}

\item{...}{not used.}
}
\value{
When \code{depth} is \code{NULL}, return the index of the maximum, \code{i}; when \code{depth} is provided, return \code{depth[i]}, the value of the depth of the maximum.
}
\description{
Find the index or depth of the maximum in a (possibly smoothed) cast. Useful to find the deep chlorophyll maximum from a fluorescence profile for example.
}
\examples{
plot(-depth ~ fluo, data=d, type="l")
# find the absolute maximum
dcm <- maxd(d$fluo, d$depth, n.smooth=0)
abline(h=-dcm, col="chartreuse4", lty="dashed")
# find the region of the maximum, along a smoothed cast
dcm <- maxd(d$fluo, d$depth, n.smooth=3, k=3)
abline(h=-dcm, col="chartreuse4")
}
