% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularise.R
\name{regularise_coord}
\alias{regularise_coord}
\title{Define an (optimised) regular coordinate}
\usage{
regularise_coord(x, start = NULL, step = NULL, control = NULL, ...)
}
\arguments{
\item{x}{vector containing the original coordinate; there are methods for \code{numeric}, \code{POSIXt} (date and times), and \code{Date} objects.}

\item{start}{initial guess for the starting point of the regular coordinate. If \code{NULL}, set to the current starting point (i.e. the minimum value of x).}

\item{step}{initial guess for the step of the regular coordinate. If \code{NULL}, set so that the new coordinate spans the entire range of \code{x}.}

\item{control}{passed as the \code{control} argument of \code{\link[stats:optim]{stats::optim()}}. \code{parscale} is already set.}
}
\description{
Based on an original coordinate, define a new one that is regular and as close as possible to the original one, overall.
}
\examples{
# Plot original and regularised coordinates
pc <- function(x, xr, ...) {
  plot(x, rep(1, length(x)), type="n",
       xlim=range(c(x, xr)), ylim=c(0.8, 1.2),
       ylab="", yaxt="n")
  abline(h=1, lty="22")
  segments(x0=x, y0=1, x1=x, y1=1.1)
  segments(x0=xr, y0=1, x1=xr, y1=0.9)
  axis(2, at=c(1.05, 0.95), labels=c("original", "regularised"))
}

# Numeric coordinate
# Already regular coordinate
x <- 1:10
pc(x, regularise_coord(x))

# Random coordinate
set.seed(1)
x <- sort(runif(10))
pc(x, regularise_coord(x))
pc(x, regularise_coord(x, step=0.05))

# Almost regular coordinate
set.seed(1)
x <- jitter(1:10, factor=0.5)
pc(x, regularise_coord(x))

# Regular coordinate with additions
x <- sort(c(1:10, 5.2))
pc(x, regularise_coord(x))

# Regular coordinate with deletions
x <- (1:10)[-c(3)]
pc(x, regularise_coord(x))

# Regular coordinate with outliers
x <- c(1:10, 12.9)
pc(x, regularise_coord(x))
x <- c(1:10, 19.9)
pc(x, regularise_coord(x))
# -> not ideal...

# Times
set.seed(1)
x <- as.POSIXct("2010-01-01 00:00:00", tz="UTC") + jitter(1:10) * 10
pc(x, regularise_coord(x))
x <- as.POSIXct("2010-01-01 00:00:00", tz="UTC") + jitter(1:10) * 3600 * 24
pc(x, regularise_coord(x))

# Dates
set.seed(1)
x <- as.Date("2010-01-01") + round(jitter(1:10)*20)
pc(x, regularise_coord(x))
}
