% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundp.R
\name{roundp}
\alias{roundp}
\title{Round values to a given precision}
\usage{
roundp(x, precision, f = round)
}
\arguments{
\item{x}{vector of values, can be numeric, POSIXct, Date.}

\item{precision}{precision at which to round, in the unit of \code{x} (for POSIXct, seconds; for Date, days).}

\item{f}{rounding function: \code{\link[base:floor]{base::floor()}}, \code{\link[base:ceiling]{base::ceiling()}}, and , \code{\link[base:round]{base::round()}}}
}
\value{
A vector of the same length as x, rounded to precision.
}
\description{
Round values to a given precision
}
\examples{
roundp(1:10, 1)
roundp(1:10, 4)
roundp(1:10, 4, floor)
roundp(1:10, 4, ceiling)

roundp(Sys.time() + 1:5, 3)
roundp(Sys.time() + 1:5, 5, floor)
# round on the hour
roundp(Sys.time(), 3600)

roundp(Sys.Date() + 1:5, 2)
roundp(Sys.Date() + 1:5, 2, floor)

# useful to bin data over a coordinate (or time)
library("dplyr")
d \%>\%
  # bin depth over 10m
  mutate(bin=roundp(depth, 10)) \%>\%
  # average each variables in each bin
  group_by(bin) \%>\% summarise_all(mean)
}
