% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratif.R
\name{stratif}
\alias{stratif}
\title{Compute an index of stratification}
\usage{
stratif(x, depth, min.depths, max.depths, n.smooth = 0, k = 2)
}
\arguments{
\item{x}{vector of the variable of interest, usually potential or in situ density.}

\item{depth}{vector of depths at which \code{x} is measured.}

\item{min.depths}{depth(s) of reference, near the surface; when \code{min.depths} is a vector, the value of x is averaged between those depths.}

\item{max.depths}{similar to \code{min.depths} but at the bottom}

\item{n.smooth}{integer, number of times to smooth the data before applying the mixed layer criteria.}
}
\value{
The value of the stratification index.
}
\description{
Compute the difference in a given variable between a two depth layers, usually the surface, mixed layer and a bottom reference layer. This is often used with density to quantify the intensity of stratification.
}
\examples{
# estimate the intensity of stratification using density
plot(-depth ~ sigma, data=d, type="l")
abline(h=-c(1, 5, 61, 65), col="red")
stratif(d$sigma, d$depth, min.depths=1:5, max.depths=61:65)

# estimate the intensity of the peak of chlorophyll at the deep
# chlorophyll maximum
plot(-depth ~ fluo, data=d, type="l")
DCM <- maxd(d$fluo, d$depth, n.smooth=2)
abline(h=-c(0, 2, DCM-2, DCM+2), col="chartreuse4")
stratif(d$fluo, d$depth, min.depths=0:2, max.depths=c(DCM-2, DCM+2))
}
\seealso{
\code{\link[=mld]{mld()}} for the computation of the depth of the mixed layer, another quantification of the intensity of mixing/stratificatin; \code{\link[=smooth]{smooth()}} for smoothing.
}
