% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ze_from_par.R
\name{ze_from_par}
\alias{ze_from_par}
\title{Detect the lower limit of the euphotic zone from PAR}
\usage{
ze_from_par(par, depth)
}
\arguments{
\item{par}{vector of Photosynthetic Active Radiation.}

\item{depth}{vector of depths at which \code{par} is measured.}
}
\value{
The depth of the lower limit of the euphotic zone.
}
\description{
Detect the lower limit of the euphotic zone from PAR
}
\details{
The lower limit of the euphotic zone (\code{Ze}) is the depth at which the downwelling PAR irradiance is reduced to 1% or its surface value. Values are accepted as "surface" until a maximum depth of 2 m.
}
\seealso{
\code{\link[=smooth]{smooth()}} for smoothing.

Other functions computing remarkable depths: \code{\link{clined}},
  \code{\link{maxd}}, \code{\link{mld}},
  \code{\link{ze_from_chla}},
  \code{\link{ze_from_surface_chla}}
}
\concept{functions computing remarkable depths}
