% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ze_from_surface_chla.R
\name{ze_from_surface_chla}
\alias{ze_from_surface_chla}
\title{Compute the lower limit of the euphotic zone from surface chlorophyll content}
\usage{
ze_from_surface_chla(chla)
}
\arguments{
\item{chla}{value(s) of chlorophyll concentration at the surface, in mg m-3, typically measured from satellite.}
}
\value{
The estimated depth of the lower limit of the euphotic zone.
}
\description{
Compute the lower limit of the euphotic zone from surface chlorophyll content
}
\details{
The lower limit of the euphotic zone (\code{Ze}) is normally computed from a profile of downwelling PAR irradiance (see \code{\link[=ze_from_par]{ze_from_par()}}). Here, it is estimated using the procedure of Lavigne et al (2012): (1) use [Chl a] to estimate Kd490, the attenuation coefficient at 490 nm (Morel and Maritorena, 2001), (2) use Kd490 to estimate KdPAR (Rochford et al, 2001), (3) use KdPAR to estimate Ze assuming an exponential decrease of light with depth.
}
\examples{
ze_from_surface_chla(0.01)
ze_from_surface_chla(0.1)
ze_from_surface_chla(1)
}
\references{
Morel and Maritorena (2001). Bio‐optical properties of oceanic waters: A reappraisal. Journal of Geophysical Research: Oceans, 106(C4), 7163-7180.

Rochford et al (2001). Importance of solar subsurface heating in ocean general circulation models. Journal of Geophysical Research: Oceans, 106(C12), 30923-30938.

Lavigne et al (2012). Towards a merged satellite and in situ fluorescence ocean chlorophyll product. Biogeosciences, 9, 2111–2125.
}
\seealso{
Other functions computing remarkable depths: \code{\link{clined}},
  \code{\link{maxd}}, \code{\link{mld}},
  \code{\link{ze_from_chla}}, \code{\link{ze_from_par}}
}
\concept{functions computing remarkable depths}
