% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clined.R
\name{clined}
\alias{clined}
\title{Detect the position of a cline}
\usage{
clined(x, depth = NULL, n.smooth = 0, k = 2)
}
\arguments{
\item{x}{vector of the variable of interest. It must have been recorded at approximately regular intervals (see \code{\link[=slide]{slide()}} as to why).}

\item{depth}{vector of depths at which \code{x} is measured; optional.}

\item{n.smooth}{integer, number of times to smooth the data before computing the moving standard deviation.}

\item{k}{order of the window; the window size is 2k+1.}
}
\value{
When \code{depth} is \code{NULL}, return the index of \code{x} corresponding to the cline, \code{i}. When \code{depth} is provided, return \code{depth[i]}, the value of the depth of the cline.
}
\description{
The cline of a variable is the depth at which it varies most sharply along the cast; here it is computed as the maximum of the moving standard deviation over the (possibly smoothed) cast.
}
\examples{
plot(-depth ~ temp, data=d, type="l")
thermocline <- clined(d$temp, d$depth)
abline(h=-thermocline, col="red")

plot(-depth ~ sal, data=d, type="l")
halocline <- clined(d$sal, d$depth)
abline(h=-halocline, col="red")
}
\seealso{
\code{\link[=slide]{slide()}} for the underlying implementation of the moving standard deviation, \code{\link[=smooth]{smooth()}} for smoothing.

Other functions computing remarkable depths: \code{\link{euphoticd}},
  \code{\link{maxd}}, \code{\link{mld}}
}
\concept{functions computing remarkable depths}
