% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\name{despike}
\alias{despike}
\title{Replace spikes by NA in a cast}
\usage{
despike(x, ...)
}
\arguments{
\item{x}{vector of data, assumed to be evenly spaced.}

\item{...}{passed to \code{\link[=spikes]{spikes()}}}
}
\description{
Replace spikes by NA in a cast
}
\seealso{
spikes
}
