% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikes.R
\name{despike}
\alias{despike}
\title{Replace spikes by NA in a cast}
\usage{
despike(x, ...)
}
\arguments{
\item{x}{vector of the variable of interest. It should be recorded at approximately regular intervals.}

\item{...}{passed to \code{\link[=spikes]{spikes()}}}
}
\description{
Replace spikes by NA in a cast
}
\examples{
x <- c(10,10,12,11,33,15,12,11,11)
plot(x, type="l")
print(x)
despike(x, k=1)
}
\seealso{
\code{\link[=spikes]{spikes()}}
}
