% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxd.R
\name{maxd}
\alias{maxd}
\title{Detect the position of the maximum}
\usage{
maxd(x, depth = NULL, n.smooth = 1, k = 2, ...)
}
\arguments{
\item{x}{vector of the variable of interest.}

\item{depth}{vector of depths at which \code{x} is measured; optional.}

\item{n.smooth}{integer, number of times to smooth the data before detecting the maximum (1 by default, because we are rarely interested in the absolute maximum, which often corresponds to a spike, but rather in the region of the maximum values).}

\item{k}{order of the moving average used for smoothing; the window size is 2k+1. NB: when data is smoothed, it should have been recorded at approximately regular intervals for the moving average to make sense.}

\item{...}{not used.}
}
\value{
When \code{depth} is \code{NULL}, return the index of \code{x} corresponding to the maximum, \code{i}. When \code{depth} is provided, return \code{depth[i]}, the value of the depth of the maximum.
}
\description{
Find the index or depth of the maximum in a (possibly smoothed) variable along a cast. Useful to find the depth of the Deep Chlorophyll Maximum (DCM) from a fluorescence profile for example.
}
\examples{
plot(-depth ~ fluo, data=d, type="l")
# find the absolute maximum
dcm <- maxd(d$fluo, d$depth, n.smooth=0)
abline(h=-dcm, col="chartreuse4", lty="dashed")
# find the region of the maximum, along a smoothed cast
dcm <- maxd(d$fluo, d$depth, n.smooth=3, k=3)
abline(h=-dcm, col="chartreuse4")
}
