% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollapply.R
\name{rollapply}
\alias{rollapply}
\title{Apply a function over a rolling window along a vector}
\usage{
rollapply(x, k, fun, n = 1, ...)
}
\arguments{
\item{x}{vector of data, assumed to be evenly spaced.}

\item{k}{order of the window; the window size is 2k+1.}

\item{fun}{function to apply in the moving window.}

\item{n}{number of times to pass the function over the data.}

\item{...}{arguments passed to \code{fun}. A usual one is \code{na.rm=TRUE}.}
}
\value{
The data passed through \code{fun}, \code{n} times.
}
\description{
Allows to compute a moving average, median, sd, mad, etc. in a generic way.
}
\examples{
# create some data and add random noise
xs <- sin(seq(0, 4*pi, length=100))
x <- xs + rnorm(length(xs), sd=0.25)
plot(x)
lines(xs)
# filter the data in various ways
# moving average
mav   <- rollapply(x, 3, mean, na.rm=T)
# running moving average
rmav  <- rollapply(x, 3, mean, na.rm=T, n=4)
# weighted moving average
wmav  <- rollapply(x, 3, weighted.mean, na.rm=T, w=c(1,2,3,4,3,2,1)/16)
# weighted running moving average
wrmav <- rollapply(x, 3, weighted.mean, na.rm=T, w=c(1,2,3,4,3,2,1)/16, n=4)
# moving median
mmed  <- rollapply(x, 3, median, na.rm=T)
lines(mav, col="red")
lines(rmav, col="red", lty="dotted")
lines(wmav, col="orange")
lines(wrmav, col="orange", lty="dotted")
lines(mmed, col="blue")
# inspect variability around filtered data
plot(rollapply(x-rmav, 7, sd))
plot(rollapply(x-mmed, 7, mad))
}
