% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Smooth a variable}
\usage{
smooth(x, k = 1, n = 1)
}
\arguments{
\item{x}{vector of the variable of interest. It must have been recorded at approximately regular intervals (see \code{\link[=slide]{slide()}} as to why).}

\item{k}{order of the window; the window size is 2k+1.}

\item{n}{number of times to smooth the data.}
}
\value{
A vector containing the smoothed variable, containing as many elements as the original variable.
}
\description{
Smooth a variable in a cast using a weighted moving average.
}
\examples{
plot(-depth ~ chla, data=d)
lines(-depth ~ smooth(chla), data=d, type="l", col="red")
lines(-depth ~ smooth(chla, k=2, n=5), data=d, type="l", col="blue")
legend(1, 0, legend=c("k=1, n=1", "k=2, n=5"), col=c("red", "blue"), lty="solid")
}
\seealso{
\code{\link[=slide]{slide()}} for the underlying implementation.
}
