% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Smooth a variable using a weighted moving average}
\usage{
smooth(x, k = 1, n = 1)
}
\arguments{
\item{x}{vector of data, assumed to be evenly spaced.}

\item{k}{order of the window; the window size is 2k+1.}

\item{n}{number of times to pass the function over the data.}
}
\value{
A vector containing the smoothed variable, containing as many elements as the original variable.
}
\description{
Smooth a variable using a weighted moving average
}
\examples{
plot(-depth ~ fluo, data=d, type="l")
lines(-depth ~ smooth(fluo), data=d, type="l", col="red")
lines(-depth ~ smooth(fluo, k=2, n=5), data=d, type="l", col="blue")
}
